refactor: removed some unpopular handlers
