+++ docs/source/modules/index.md
@@ -14,7 +14,7 @@ medias
 regex
 stickers
 translator
-users_groups
+users
 web
 ```
 
@@ -46,5 +46,5 @@ Below is a list of all available modules in _PyKorone_:
 - [Regex](./regex): Test regular expressions.
 - [Stickers](./stickers): Get stickers and steal stickers to a sticker pack.
 - [Translator](./translator): Translate text to different languages using DeepL.
-- [Users & Groups](./users_groups): Fetches information about users and groups.
+- [Users](./users): Fetches information about Telegram and _PyKorone_ users.
 - [Web Tools](./web): Get information about IP addresses and domains.
