+++ docs/source/modules/users.md
@@ -1,10 +1,9 @@
 # Users & Groups
 
-This module allows retrieval of public information that are available trough the Telegram API about users and groups.
+With this module you can get information about bot users, remove deleted accounts from groups, and more.
 
 ## Commands
 
 - `/user (user)`: Returns information about a user. If no user is specified, it returns information about the user who sent the command.
-- `/group (group)`: Returns information about a group. If no group is specified, it returns information about the group where the command was sent.
 - `/id`: Returns the ID of the user who sent the command, or the group to which the command was sent. Can also be used to get a user's ID by replying to a message from that user.
 - `/cleanup`: Removes deleted accounts from the group where the command was sent.
